+++ classes/model/manager/feedback_manager.php
@@ -69,7 +69,12 @@ class feedback_manager extends manager {
         $this->dao->update($feedback);
     }
 
-    public function lock_feedbacks_by_version(version $version) {
+    public function lock_feedbacks_by_version(version $version, $userid) {
+
+		if (!sequentialdocuments_has_lock_feedback_rights($this->instanceid, $userid)) {
+			throw new unauthorized_access_exception(1006, 'mod_sequentialdocuments');
+		}
+
         $feedbacks = $this->dao->get_all_entities_where(array('versionid' => $version->get_id()));
         if ($feedbacks !== false) {
             foreach ($feedbacks as $feedback) {
@@ -100,7 +105,7 @@ class feedback_manager extends manager {
         return $this->dao->get_all_entities_where(array('versionid' => $id));
     }
 
-    public function delete_feedback($feedbackid) {
+    public function delete_feedback($feedbackid, $userid) {
         $feedback = $this->dao->get_entity($feedbackid);
 
         if (!($feedback instanceof feedback)) {
@@ -109,10 +114,20 @@ class feedback_manager extends manager {
             throw new unauthorized_access_exception(1000, 'mod_sequencialdocuments');
         }
 
+		if (!sequentialdocuments_has_feedback_suppression_rights($this->instanceid, $userid)) {
+			throw new unauthorized_access_exception(1003, 'mod_sequentialdocuments');
+		}
+
         $this->dao->delete($feedback);
     }
 
-    public function delete_feedbacks_by_versionid($versionid) {
+    public function delete_feedbacks_by_versionid($versionid, $userid) {
+
+		if (!sequentialdocuments_has_version_suppression_rights($this->instanceid, $userid) ||
+			!sequentialdocuments_has_feedback_suppression_rights($this->instanceid, $userid)) {
+			throw new unauthorized_access_exception(1003, 'mod_sequentialdocuments');
+		}
+
         $feedbacks = $this->dao->get_all_entities_where(array('versionid' => $versionid));
         if ($feedbacks !== false) {
             foreach ($feedbacks as $feedback) {
@@ -123,26 +138,41 @@ class feedback_manager extends manager {
 
     public function get_feedback_html_by_id(
                                             $id,
+											$userid,
                                             document_manager $documentmanager,
                                             version_manager $versionmanager,
                                             $contextid) {
 
+		if (!sequentialdocuments_has_feedback_read_rights($this->instanceid, $userid)) {
+			return '';
+		}
+
         return $this->get_feedback_html_by_feedback_instance(
-                $this->dao->get_entity($id), $documentmanager, $versionmanager, $contextid
+							$this->dao->get_entity($id),
+							$userid,
+							$documentmanager,
+							$versionmanager,
+							$contextid
         );
     }
 
     public function get_feedback_html_by_feedback_instance(
                                                             feedback $feedback,
+															$userid,
                                                             document_manager $documentmanager,
                                                             version_manager $versionmanager,
                                                             $contextid) {
         if ($feedback === null) {
             throw InvalidArgumentException();
         }
+
         if ($feedback->get_instanceid() != $this->instanceid) {
             throw new unauthorized_access_exception(1000, 'mod_sequencialdocuments');
         }
+
+		if (!sequentialdocuments_has_feedback_read_rights($this->instanceid, $userid)) {
+			return '';
+		}
 
         $setnewversionlink = '';
         $filelist = '';
@@ -174,15 +204,25 @@ class feedback_manager extends manager {
 
             // TODO : implement the version creation from feedback function.
             $version = $versionmanager->get_version($feedback->get_versionid());
-            if (!$version->is_locked() && $versionmanager->is_last_version($documentmanager, $version)) {
+            if (!$version->is_locked() && 
+					$versionmanager->is_last_version($documentmanager, $version) &&
+					sequentialdocuments_has_version_creation_rights($this->instanceid, $userid)) {
                 $filelist .= '<br /><br /><a href="#">Create new version from this feedback</a>';
             }
         }
 
         $links = '';
         if (!$feedback->is_locked()) {
-            $links = '<a href="'.get_update_feedback_url($feedback->get_id(), $this->instanceid).'">Edit</a> '.
-                    '<a href="'.get_delete_feedback_url($feedback->get_id(), $this->instanceid).'">Delete</a>';
+			if (sequentialdocuments_has_feedback_edit_rights($this->instanceid, $userid)) {
+				$links .= '<a href="'.
+							get_update_feedback_url($feedback->get_id(), $this->instanceid).
+							'">Edit</a> ';
+			}
+			if (sequentialdocuments_has_feedback_suppression_rights($this->instanceid, $userid)) {
+				$links .= '<a href="'.
+							get_delete_feedback_url($feedback->get_id(), $this->instanceid).
+							'">Delete</a>';
+			}
         }
 
         $links = '<aside class="sqds-bottom-right">'.
