+++ classes/model/manager/version_manager.php
@@ -97,6 +97,7 @@ class version_manager extends manager {
 
     public function delete_version(
                                     $versionid,
+									$userid,
                                     document_manager $documentmanager,
                                     feedback_manager $feedbackmanager) {
 
@@ -108,19 +109,33 @@ class version_manager extends manager {
             throw new unauthorized_access_exception(1000, 'mod_sequencialdocuments');
         }
 
+		if (!sequentialdocuments_has_version_suppression_rights($this->instanceid, $userid)) {
+			throw new unauthorized_access_exception(1002, 'mod_sequentialdocuments');
+		}
+
         if ($this->is_last_version($documentmanager, $version)) {
             $documentmanager->notify_version_deletion($this, $version);
         }
-        $feedbackmanager->delete_feedbacks_by_versionid($versionid);
+        $feedbackmanager->delete_feedbacks_by_versionid($versionid, $userid);
         $this->dao->delete($version);
     }
 
-    public function lock(version $version) {
+    public function lock(version $version, $userid) {
+
+		if (!sequentialdocuments_has_lock_version_rights($this->instanceid, $userid)) {
+			throw new unauthorized_access_exception(1005, 'mod_sequentialdocuments');
+		}
+
         $version->set_locked(true);
         $this->dao->update($version);
     }
 
-    public function unlock(version $version) {
+    public function unlock(version $version, $userid) {
+
+		if (!sequentialdocuments_has_lock_version_rights($this->instanceid, $userid)) {
+			throw new unauthorized_access_exception(1005, 'mod_sequentialdocuments');
+		}
+
         $version->set_locked(false);
         $this->dao->update($version);
     }
@@ -152,11 +167,17 @@ class version_manager extends manager {
         return $this->dao->get_all_entities_where(array('documentid' => $id));
     }
 
-    public function delete_versions_by_documentid($documentid, $feedbackmanager) {
+    public function delete_versions_by_documentid($documentid, $userid, $feedbackmanager) {
+
+		if (!sequentialdocuments_has_document_suppression_rights($this->instanceid, $userid) ||
+			!sequentialdocuments_has_version_suppression_rights($this->instanceid, $userid)) {
+			throw new unauthorized_access_exception(1002, 'mod_sequentialdocuments');
+		}
+
         $versions = $this->dao->get_all_entities_where(array('documentid' => $documentid));
         if ($versions !== false) {
             foreach($versions as $version) {
-                $feedbackmanager->delete_feedbacks_by_versionid($version->get_id());
+                $feedbackmanager->delete_feedbacks_by_versionid($version->get_id(), $userid);
                 $this->dao->delete($version);
             }
         }
@@ -168,17 +189,27 @@ class version_manager extends manager {
 
     public function get_version_html_by_id(
                                             $id,
+											$userid,
                                             document_manager $documentmanager,
                                             feedback_manager $feedbackmanager,
                                             $contextid) {
 
+		if (!sequentialdocuments_has_version_read_rights($this->instanceid, $userid)) {
+			return '';
+		}
+
         return $this->get_version_html_by_version_instance(
-                        $this->dao->get_entity($id), $documentmanager, $feedbackmanager, $contextid
+							$this->dao->get_entity($id),
+							$userid,
+							$documentmanager,
+							$feedbackmanager,
+							$contextid
         );
     }
 
     public function get_version_html_by_version_instance(
                                                             version $version,
+															$userid,
                                                             document_manager $documentmanager,
                                                             feedback_manager $feedbackmanager,
                                                             $contextid) {
@@ -191,13 +222,17 @@ class version_manager extends manager {
             throw new unauthorized_access_exception(1000, 'mod_sequencialdocuments');
         }
 
+		if (!sequentialdocuments_has_version_read_rights($this->instanceid, $userid)) {
+			return '';
+		}
+
         $html = '';
         $feedbacks = $feedbackmanager->get_feedbacks_by_versionid($version->get_id());
 
         if (count($feedbacks) > 1 || (count($feedbacks) == 1 && $feedbacks[0] != false)) {
             foreach ($feedbacks as $feedback) {
                 $html .= $feedbackmanager->get_feedback_html_by_feedback_instance(
-                                            $feedback, $documentmanager, $this, $contextid
+                                            $feedback, $userid, $documentmanager, $this, $contextid
                 );
             }
         }
@@ -227,15 +262,19 @@ class version_manager extends manager {
         $editlink = '';
         $postfeedbacklink = '';
         if (!$version->is_locked()) {
-            $editlink =
-                    '<a href="'.get_update_version_url($version->get_id(), $version->get_instanceid()).'">'.
-                        'Edit'.
-                    '</a>';
+			if (sequentialdocuments_has_version_edit_rights($this->instanceid, $userid)) {
+				$editlink =
+						'<a href="'.get_update_version_url($version->get_id(), $version->get_instanceid()).'">'.
+							'Edit'.
+						'</a>';
+			}
 
-            $postfeedbacklink =
-                    '<a href="'.get_add_feedback_url($version->get_id(), $version->get_instanceid()).'">'.
-                        'Post a feedback'.
-                    '</a> ';
+			if (sequentialdocuments_has_feedback_creation_rights($this->instanceid, $userid)) {
+				$postfeedbacklink =
+						'<a href="'.get_add_feedback_url($version->get_id(), $version->get_instanceid()).'">'.
+							'Post a feedback'.
+						'</a> ';
+			}
         }
 
         $versionid = $version->get_id();
